move security advisor to new directories (#261)
