+++ stix_shifter_modules/security_advisor/stix_translation/README.md
