+++ stix_shifter_modules/security_advisor/stix_translation/json/to_stix_map.json
