+++ stix_shifter_modules/security_advisor/stix_transmission/security_advisor_auth.py
