+++ stix_shifter_modules/security_advisor/stix_transmission/security_advisor_error_mapper.py
@@ -1,5 +1,5 @@
-from .....utils.error_mapper_base import ErrorMapperBase
-from .....utils.error_response import ErrorCode
+from stix_shifter_utils.utils.error_mapper_base import ErrorMapperBase
+from stix_shifter_utils.utils.error_response import ErrorCode
 
 error_mapping = {
         'service_not_availiable' :  ErrorCode.TRANSMISSION_CONNECT,
