+++ stix_shifter_modules/security_advisor/stix_transmission/security_advisor_results_connector.py
@@ -1,6 +1,6 @@
-from ..base.base_results_connector import BaseResultsConnector
+from stix_shifter_utils.modules.base.stix_transmission.base_results_connector import BaseResultsConnector
 from .utils.stix_pattern_processor import StixPatternProcessor
-from .....utils.error_response import ErrorResponder
+from stix_shifter_utils.utils.error_response import ErrorResponder
 from .security_advisor_auth import SecurityAdvisorAuth
