+++ stix_shifter_modules/security_advisor/stix_transmission/security_advisor_status_connector.py
@@ -1,5 +1,6 @@
-from ..base.base_status_connector import BaseStatusConnector
-from ..base.base_status_connector import Status
+from stix_shifter_utils.modules.base.stix_transmission.base_status_connector import BaseStatusConnector
+from stix_shifter_utils.modules.base.stix_transmission.base_status_connector import Status
+
 
 class SecurityAdvisorStatusConnector(BaseStatusConnector):
     def __init__(self, host, auth):
