+++ stix_shifter_modules/security_advisor/stix_transmission/utils/sa_findings_api.py
