+++ stix_shifter_modules/security_advisor/stix_transmission/utils/sa_occurence_finder.py
