+++ stix_shifter_modules/security_advisor/stix_transmission/utils/statement_parser.py
