+++ stix_shifter_modules/security_advisor/stix_transmission/utils/stix_pattern_processor.py
